Extend SVD procedure to DM Pulay mixing
The implicit inversion in the DIIS procedure is carried out using
SVD techniques if 

UseSVDInPulay T

The tolerance to discard small singular values is hardwired to 1.e-6
('rcond' in the sense of the LAPACK routine dgelss).

Rank degradation history information might be used to determine
the appropriateness of "kicks" (to be implemented).

(+ The 'solve_with_svd' routine is in the new m_svd.F90 module file)

added:
  Src/m_svd.F90
modified:
  Src/Makefile
  Src/m_new_dm.F90
  Src/m_pulay.F90