Implement handling of flag file '0_NORMAL_EXIT' for jobList
The file signals the successful end of execution of a Siesta run.  It
is removed at the beginning, if it exists, and 'touched' at the end of
the run.  The 'jobList' suite checks for the existence of this file
(falling back to the old scheme of checking for the .EIG file) to
decide whether to (re)submit the job.

For older versions of Siesta, the same functionality could perhaps be achieved
at the level of the shell.

modified:
  Src/siesta_end.F
  Src/siesta_init.F
  Util/JobList/Src/jobList.f90