Increased parameter minloc in rhoofd and vmat from 100 to 1000 for efficiency
Parameter minloc fixes the size of a buffer array, to hold matrix elements
of the density matrix or the hamiltonian, between orbitals that overlap in
a certain region. Present memory sizes allow to increase it, diminishing
the frequency of copies from the sparse-format arrays. The changes in
mesh1d and atmfuncs are unconsequential.

modified:
  Src/SiestaXC/mesh1d.f90
  Src/atmfuncs.f
  Src/rhoofd.F
  Src/vmat.F