Smarter interface with interval guessing by inertia counts
This is the implementation by Georg Huhs of the new interface.

There are now two entry points to PEXSI:

f_ppexsi_inertiacount_interface
f_ppexsi_solve_interface

together with new parameters to determine the behavior of the code.
Full documentation will be produced at the time of merging.

(the dummy file pexsi_interface.F90, and the Makefile, was updated
by A. Garcia)

removed:
  Src/pexsi.h
added:
  Docs/pexsi-gh-richer.CHANGES
  Src/pexsi_inertia.h
  Src/pexsi_solve.h
modified:
  Src/Makefile
  Src/m_pexsi_solver.F90
  Src/pexsi_interface.F90