Move wannier analysis to 'siesta_analysis'. Reinstate call to siesta_end
A previous patch had removed the call to siesta_end in the main program.
It has been re-instated. 

The call to siesta2wannier90 has been moved to 'siesta_analysis'. In
case the user requests a local_DOS calculation (which destroys the scf
DM), the new call is placed just before that section.

modified:
  Src/Makefile
  Src/siesta.F
  Src/siesta_analysis.F