Lift the restriction on multiple l-shells in atom
ATOM can now work with several valence shells of the same angular
momentum (for example, 5s and 6s in Ba, when treating 5s as semicore).

There is no need to use ionized configurations for these kinds of
problems. ATOM will pseudize the state with the lowest n for a given
l, so the input format does not change.

added:
  Pseudo/atom/Tutorial/PS_Generation/Ba/Ba.multiple.shells.inp
modified:
  Pseudo/atom/pseudo.f
  Pseudo/atom/wrapup.f