Optionally read the Hamiltonian for analysis
With the option

   read-H-from-file T

the program will attempt to read H from the file "H_IN". This is
useful when performing analyses which only need H (and S), but not
the DM. 

For example, to compute bands, or (p)DOS:

(... options for bands or (p)DOS...)
max-scf-iterations 0
read-H-from-file T

Optionally, since the Fermi level is not computed in these cases, it
can be set directly in the fdf file:

   manual-fermi-level -3.05 eV

The program might still attempt to initialize the DM, and print messages
mentioning DM input. These are cosmetic issues, but the user should be
aware that this options are to be used with care.


modified:
  Src/siesta_analysis.F