Disable parallel timer by default
When large numbers of processors are used, the "parallel" timer
(producing the TIMES file) might be inefficient, and introduce
inbalances in the "tree_timer" (based on walltime on the root node).

The "parallel" timer has been disabled by default. It can be 
enabled by

  Use.Parallel.Timer T

(The "tree_timer" is on by default. It can be disabled by

  Use.Tree.Timer F
)

modified:
  Src/siesta_init.F
  Src/timer.F90