Added real solver based on the MRRR method in Scalapack.
This method allows to compute only a subset of eigenpairs.

First version with query for sizing of work arrays.
It could probably be optimized for memory in Z, since
only a subset of eigenvectors is dealt with.

Set the option with

 Diag.MRRR T

and specify number of eigenstates to compute by

NumberOfEigenstates N

(A. Garcia)

added:
  Src/rdiag_mrrr.F
modified:
  Src/Makefile
  Src/diagg.F