Added functionality for tracing with extrae.
For each timing event a user function with the same name is registered. 
At the end of the run the file 'siesta_user_labels.pcf' containing the labels needed for later analysis is created. The content of this file needs to be appended to the *.pcf file generated by the tracing. 

For activating this functionality the code has to be compiled with the option 
  -DTRACING
and linked with the extrae-library. 

The file Src/extrae_module.f90 is taken from extrae 2.4 and might need to be updated when using later versions of extrae.