Bugfix in Makefile
Extended functionality for tracing with extrae.

When compiling with 
  -DTRACING_SOLVEONLY
the tracing is deactivated after initializing Siesta and activated again before starting the first pexsi-solve phase. 
This ensures that for some time the pexsi-solve is not corrupted by tracing-IO. If this time-window is too short, the event buffer size can be increased in extrae.xml

The user labels are now written with increasing numbering.