Document the PEXSI features and update interface to v0.7.3
Wrote the first version of a new section of the manual devoted to
PEXSI.

Updated the interpretation of the PEXSI.safe-width-ef-solver parameter
to make it consistent with the psimilar parameter for
inertia-counting.

Deprecated the use of the parameter PEXSI.inertia-count.

A minor cosmetic change in f_ppexsi_interface.f90 in the official
v0.7.3.

modified:
  Docs/siesta.tex
  Src/f_ppexsi_interface.f90
  Src/m_pexsi_driver.F90
  Src/Sys/macosx-pexsi.make