Allow Siesta to read KB projectors from a PSML file
First implementation, without touching the premilinary "basis_specs"
stage (i.e., the user might specify KB options in the fdf file, but
they will be overridden, except lmxkb, by the contents of the
<projectors> element in the PSML file).

To enable this option meaningfully, the user must specify:

use-psml-vlocal T
use-psml-kb-projectors T

and there must be a SystemLabel.psml file in the working directory.

Note that the PAOs are still generated by the standard procedure
involving the semilocal potentials, which might be read from the same
PSML file, or even from another pseudopotential file. This has to be
made more robust and flexible.

Upon reading the local pseudopotential, the program must construct the
associated "local charge density". It does this using routines in the
module m_vlocal_chlocal. To enable some debugging checks, some options
formerly hardwired in this suite of routines have been made available
to the user:

	Vlocal-from-Chlocal  T

will force the use of the "small core" method of construction of
Vlocal: generation of a localized pseudo-gaussian charge density
followed by integration of the Poisson equation. This tends to give
smooth charges (by construction) and a smooth Vlocal. The alternative
method, used for large-core situations, is a direct fit of Vlocal to
match the Vl's beyond the range at which they are equal but not as far
as the point at which they become equal to -Z/r. The fit is by default
done maintaining the continuity of the second derivative, but the user
can request the continuity of the third derivative by using the option

        Vlocal.Fit.3derivs  T

Note that this "fitting" method might lead to Vlocals with subtle
shapes that manifest themselves more clearly in the form of
chlocal. It is advisable to look at plots of Chlocal (the data is now
available by default) to diagnose this effect properly. Note that when
importing Vlocal from a "foreign" PSML file, the shape of Chlocal
might not resemble a pseudogaussian at all. This is particularly
noticeable when using core-corrections and high-order continuity of
the fit involved in the construction of Vlocal, as in D.R. Hamann's
program oncvpsp.

Some operational options in 'atom.F' have changed defaults. The
following are true by default:

    write_ion_plot_files = fdf_boolean('WriteIonPlotFiles',.true.)
    debug_kb_generation  = fdf_boolean('Atom.Debug.KB.Generation',.true.)
    new_kb_reference_orbitals  = fdf_boolean('KB.New.Reference.Orbitals',.true.)

Note in particular that the last one will turn on the new Rmax and
nrwf defaults for the construction of KB projectors, as implemented by
J. Junquera.

An important bug fix in psop: the projectors are handled in the Siesta
tables divided by r**(l+1). The PSML output produced by psop now
contains the standard form, without the factor.  (A further, now
irrelevant, fix involved the removal of a spurious factor of 0.5 in
the old (wrong) projector output).

The code for psop and that in atom.F and m_vlocal_chlocal.f should be merged
for clarity, and to avoid the use of slightly different options between a straight
Siesta calculation and the generation of a PSML file.

modified:
  Src/atom.F
  Src/atom_options.F90
  Src/m_vlocal_chlocal.f
  Util/Pseudo-operator/module_pseudo.F