Control MPI timing from fdf file. Choose wall- or cpu-time
By simply turning 'timer_mpi' into an external function (defined in
timer.F90), the user can control the generation of timing information
from MPI calls. The relevant fdf line is

    time-mpi-calls T

This option can work with both the 'tree-timer' (use-tree-timer T),
and the traditional 'linear' timer. In the former case only the root
node information is printed. For the latter, load-balance information
is also printed. 

The kind of timing performed can be controlled with the fdf idiom

   use-wall-time T    

If .true., the wall-time is used, otherwise calls to the 'cpu_time'
intrinsic are used. Wall-time is used by default for the tree-timer,
whereas the default for the 'linear' timer is to use the cpu time.

removed:
  Src/MPI/timer_mpi.F90
modified:
  Src/MPI/Makefile
  Src/MPI/mpi__type_s.f90
  Src/MPI/mpi__type_sv.f90
  Src/MPI/mpi__type_v.f90
  Src/MPI/mpi__type_vs.f90
  Src/MPI/mpi_siesta.f90
  Src/m_timer.F90
  Src/siesta_init.F
  Src/timer.F90