Added Docs/release.sh script for easy creation of tar.gz
This script enables easier release tar.gz files with a consistent
format.
The main functionality is that it will create a folder:
   siesta-releases
in the main folder where it will create a branch at the specified
tag. It will subsequently build the documentation and create
any necessary changes to the code that is necessary before
making the code available publicly.

This script is provided by Nick R. Papior