Temporary work-around for supercell changes
The TS branch introduced a kill of siesta when the supercell
was changed during an MD. This may not always be the case if the
supercell structure is correct, but the connected supercells are
changing. In that case will it be acceptable to continue. However,
there may be cases where this is not applicable.
This will be fully fixed in future cases where the supercell information
is retained in restart files.