Added print-out of maximum force difference in grdsam
This enables one to more easily figure out if the grid
sampling is required or not.
siesta now also prints the standard-deviation.
We have found that in certain systems (a Cu-Ni slab-alloy)
the forces were highly unsymmetric (where they should have
been symmetric). The solution was to grid-cell-sample
along the normal vector to the surface. This helped
alot and fixed the problem, however, before doing a
fine-mesh sampling it can be nice to get a feel for it
using very few points.