Enabled reading of DM with different spin-components
This is handy for doing large scale calculations with
multiple spins. First start by calculating the non-polarized
case, then increase spin to polarized, then increase to
non-collinear, and finally to spin-orbit. This has been
tested on a very simple system to check it works.
 siesta -fdf Spin:none run.fdf
 siesta -fdf Spin:polarized run.fdf
 ...
will now work.
This partially fixes lp:1645388.

The reading/writing of DM has changed to not take into account the
number of spin-components. Thus checks for spin should be taken care
of after reading DM (for TS and S).