Made transiesta and tbtrans quit if the transfer matrix is 0
If the user supplies an electrode where the transfer matrix is 0
transiesta and tbtrans will now die, regardless of whether the user
wants this.
In fact, we could allow this by treating it as a molecular end, however
it will then not be a reservoir.
This sanity check should prevent users from doing wrong calculations.

Also fixed print-out of electrode names if options are inconsistent in
m_ts_electype.