Fixed coordinate exchanges to LUA and added a bunch more
In the LUA hook one now has access to the unit conversion
table:

  siesta.Units

which contains all the unit conversions used internally in
siesta.

Added all the tolerance variables to the siesta dictionaries
for communicating with lua

A bug was found for communicating the coordinates from siesta
to lua. When a supercell calculation was performed the xa(:,1:na_s)
was sent, while it should only be xa(:,1:na_u). This has been fixed.

Minor fixes in the usage of the options in siesta_options and read_options
The default variable was not really used, and has been removed.