Almost finalized the flook interoperability
Added documentation of the Lua scripting capability.
In the documentation I have added a link to the flos library
which already implements a few enhancements over the previous things.

Added a new state in Lua, INIT_MD which is required to be able to
do MeshCutoff convergences, this required the mesh cut-off to be retrieved
from setup_H0.

Added target attribute to the siesta-dictionary to ensure they are passed by
reference.

Added Mesh.Cutoff.Minimum and Mesh.Cutoff.Used.

Moved the SCF_LOOP state to be just after mixing has been performed.

Added the new states to the h2o_lua example.