New method to determine chlocal
It is based in the calculation of the laplacian of Vlocal by finite
differences, using a routine by Don Hamann, and subsequent direct
interpolation into the linear tables used by Siesta with another
Hamann routine.

The weak point is still the determination of the chlocal cutoff. The
use of a Hamann-style grid forces a change in the order in which the
selection of the final radial point and the normalization is carried
out. For full compatibility with Siesta a chlocal computed directly
on the Froyen grid will be needed (next commit).

added:
  Src/hamann.f90
modified:
  Src/Makefile
  Src/atom.F