Change of mix_scf_first based on DM input
Based on a couple of cases where the DM would converge but would
have some memory of the initial (previous) DM, we have decided to
only allow mixing the first SCF step IFF the DM is filled with
atomic data, or if the read in DM has a different charge than it
is supposed to.
This should yield a better converged final DM and thus, forces.