Initializing the off-site Spin-Orbit implementation within trunk-604
- The implementation will be done within the trunk-604
and the following updates will be renamed as trunk-604-off-site-#,
where # will depict different updates.

- Adding some necessary variables and flags within m_spin.F90
module.

- Included a new output file called "SystemLabel.offSO" to 
debug the implementation. The option is only available for 
the off-site and the flag to add to the fdf file is: "Spin SO+deb". 
This file will have information regarding the progress of the 
calculation as well as, for example, the hermicity checks, 
the evolution of the total magnetic moment and its orientation 
w.r.t. X, Y and Z axis, and many more.     

- Added the off-site SO proper citation to m_cite.F90  

- Added Enl_SO energy variable to the m_energies.F90 module. 
This will be the off-site SO energy contribution to the total 
SC energy.  

- It is possible to follow the changes made related with the 
implementation along the code just searching for "CC RC". 


modified:
  .bzrignore
  Src/basis_specs.f
  Src/basis_types.f
  Src/initatom.f
  Src/m_cite.F90
  Src/m_energies.F90
  Src/m_spin.F90
  Src/siesta_init.F
  version.info