Work around reshape in mesh3d for SiestaXC
This fixes: lp:1665294

Instead of doing reshape, we implement direct loops with
an interface for copying, and one for adding.
This should limit the heap memory consumption.

I have tested this on 10 of the tests in the test directory,
with no changes (however, only on 2 cores).
Before releasing 4.0.1 we should rerun all tests.