Implement tion of the Kleinman-Bylander operator to support the off-site SO
- Added two new subroutines within sparse_matrices.F. The 
first one depicts the magnetic moments and its orientation: 
MM_HDM('DM/H'). The second one checks the hermicity of H or 
DM: herm_HDM( 'DM/H' ). Both subroutines can be called passing
H or DM. Currently, it can be used in serial mode (in principle 
is for debugging, maybe in the future it will be useful using 
during an usual calculation in order to check if the orientation 
of the magnetization is deviated from its original aligment, etc.)

- The subroutines comKB, KBproj, KBgen and nlefsm have kept 
as they are and will be used for any type of calculation not 
related with the off-site SO. New specific subroutines have been 
created for the off-site SO such as comKB_SO, KBgen_SO, KBproj_SO 
and nlefsm_SO. In addition, some subroutines within radial module 
and in register_rfs, have been duplicated and adapted to the 
off-site SO. The names as usual are ended in "_SO". 

- One subroutine has been created in order to obtain the 
Clebsh-Jordan prefactors coefficients and the non-local 
part of the KB operator when the off-site SO is included: 
cal_Vj_SO. This subroutine is inside nlefsm module.  

- It has been updated the new gindex ordenation method for 
the new KBs in the off-site SO. So new_matel is called with 
the proper KB and orbital gindex giving the proper Ski integrals.  

- The current status of the implementation is the calculation 
of the non-selfconsistent part of the hamiltonian, i.e., 
up to setup_H0.

- It has been tested as well that the current features of 
SIESTA are kept working such as the on-site SO implementation.

- So far (up to setup_H0) it works in serial and parallel modes.   

- Some bug fixes. 


modified:
  .bzrignore
  Src/atm_transfer.f
  Src/atm_types.f
  Src/atmparams.f
  Src/atom.F
  Src/basis_io.F
  Src/basis_specs.f
  Src/basis_types.f
  Src/broadcast_basis.F
  Src/initatom.f
  Src/m_new_dm.F90
  Src/m_spin.F90
  Src/matel_registry.F90
  Src/nlefsm.f
  Src/old_atmfuncs.f
  Src/radial.f
  Src/register_rfs.F90
  Src/setup_H0.F
  Src/setup_hamiltonian.F
  Src/siesta.F
  Src/siesta_forces.F90
  Src/siesta_init.F
  Src/sparse_matrices.F
  Src/state_init.F