Fixed memory leaks in fdict, and transiesta related routines
Several memory leaks was detected using valgrind, nearly all
have been fixed.

Currently, there is still a memory leak regarding the dictionary
when concatenating. However, it is 2 bytes. I will amend a fix for this
in a later revision (track fdict on github for when the fix will be
available).