Fully implementation of the off-site SO (serial runs)
- Folowing with the previous revision of the present 
off-site SO branch (i.e., trunk-604-off-site, rev. 606), 
I have changed setup_hamiltonian.F, dhscf.f, diag3g.F 
and diag3k.f and many more. diag3g and diag3k subroutines 
are only used when SO is included within gamma point 
calculation or bulk systems, respectively. In addition, I 
added to the total hamiltonian the H0_SO matrix elements. 
Regarding vmat and rhoofd subroutines, it has been included 
in their modules vmat_SO and rhoofd_SO, initially created 
for the off-site SO, however, vmat and rhoofd have been modied 
for the off-site SO accordigly and they work fine as they are. 
So it is not necessary the use of those subroutines ended in 
'_SO'.   

- In SiestaXC/ggaxc.f it has been changed the current 
non-collinear scheme for the old one used previously in 
Src/xc.f. In doing so it is possible to use the current 
implemented vdW functionals. 

- In the present version is used the transpose matrices. This 
has to be fixed for non-transpose "formalism". The transpose 
indexes are calculated within state_init subroutine and to 
use them it has to be called everywhere as indt = listht(ind). 
This version of the branch won't work in parallel because of the 
transpose matrices... So this has to disappear in the following 
revions of the branch.      

- LDA+U has to be checked as well. So far it is not tested 
both features at the same time.  

- Some comments have left along the code in order to check 
possible future bugs and/or new features related with the 
off-site SO implementation. These will be removed in the future
as well.  

- So far the tests done are Pt2 and FePt-L1_0 MAEs calculation
and some bands calculation. 

- Note: In the fdf file has to be written the following options
for the off-site SO (mandatory):

    * PAO.OldStylePolOrbs F
    * Restricted.Radial.Grid  F
    * DM.MixSCF1  F


modified:
  .bzrignore
  Src/SiestaXC/ggaxc.f
  Src/atmfuncs.f
  Src/atom.F
  Src/bands.F
  Src/compute_dm.F
  Src/compute_energies.F90
  Src/dfscf.f
  Src/dhscf.F
  Src/diag3g.F
  Src/diag3k.F
  Src/diagon.F
  Src/final_H_f_stress.F
  Src/listsc.f
  Src/m_energies.F90
  Src/m_new_dm.F90
  Src/m_pulay.F90
  Src/mixer.F
  Src/nlefsm.f
  Src/pdos3g.F
  Src/pdos3k.F
  Src/rhoofd.F90
  Src/scfconvergence_test.F
  Src/setup_H0.F
  Src/setup_hamiltonian.F
  Src/siesta_forces.F90
  Src/sparse_matrices.F
  Src/state_init.F
  Src/vmat.F90
  Src/write_subs.F
  version.info